Changes to support Windows builds
Replaces direct use of /tmp, to fix Windows issues

- Add ScratchDir/File methods to CloudEnvImpl
- Fix the ObjectPath to always be set to unix-style directory
- Use RocksDB DestroyDir rather than custom implementation
- Change the CopyToFromS3 test to append smaller blocks.  Windows barfs when appending large blocks to files (RocksDB, not Cloud issue?)

+++ cloud/aws/aws_retry.cc
@@ -125,7 +125,7 @@ Status AwsCloudOptions::GetClientConfiguration(
   // Setup how retries need to be done
   config->retryStrategy = std::make_shared<AwsRetryStrategy>(env);
   if (cloud_env_options.request_timeout_ms != 0) {
-    config->requestTimeoutMs = cloud_env_options.request_timeout_ms;
+    config->requestTimeoutMs = static_cast<long>(cloud_env_options.request_timeout_ms);
   }
 
   config->region = ToAwsString(region);
